Drop the maximum recommended log size from 16MiB to 8MiB.
Now we compress logs, 16MiB is a *lot* of logging, often over 100MiB
when decompressed. This can cause trouble for tools (and their
corresponding timeouts) if they're expecting 16MiB to mean "less than
16MiB" rather than "about 10 times 16MiB".

This patch doesn't address the issue of devices that are already
configured too high, since it's not obvious that we reasonably can. A
user can set the system property themselves, and we have no way to
distinguish the two cases.

Bug: https://issuetracker.google.com/176959217
Test: treehugger
Change-Id: Iba92061b278c487c3dc9a98aa54f014dfb2ea1c9
Signed-off-by: Akash Srivastava <akashniki@gmail.com>
Signed-off-by: Joey Huab <joey@evolution-x.org>